Improve the speed of the intersection algorithm up by the factor of 8 (branchless and vectorizes really well). Small settings system changes.

+++ tests/shmemalloc.cpp
@@ -59,7 +59,7 @@ int main()
 				sem.Post();
 			}
 
-			for (int i = 0; i < vec->size(); i++) {
+			for (size_t i = 0; i < vec->size(); i++) {
 				printf("%d\n", (*vec)[i]);
 			}
 		} else {
